+++ core/include/booker/io/DirectoryUtils.hpp
@@ -1,0 +1,8 @@
+#ifndef DIRECTORYUTILS_HPP
+#define DIRECTORYUTILS_HPP
+
+#include <filesystem>
+
+bool flattenDirectory(std::filesystem::path const& dirPath);
+
+#endif // DIRECTORYUTILS_HPP