+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- io directory in core
+- TempDirectory class to manage a temporary directory
+- DirectoryUtils files to get utilities functions
+
 ## [0.0.8] - 2026-05-11
 
 ### Added